refactor(benchmark): split scorer.py into scoring_primitives + grouping modules
Zero behavior change, pure code motion. Preparation for the mech-analytics
consumer migration (PR #11 §6): analyze.py rewrite in a subsequent commit
will import accumulate_row from grouping without pulling in scorer's CLI /
rebuild / sweep surface.

Layout:

- benchmark/scoring_primitives.py (new) — constants (MIN_SAMPLE_SIZE,
  CALIBRATION_BINS, DIFFICULTY_THRESHOLDS, etc), single-row score
  functions (brier_score, edge_score, log_loss_score) and classifiers
  (classify_difficulty, classify_liquidity, classify_disagreement,
  classify_horizon, disagree_bucket, _bin_label, _empty_group,
  _derive_diagnostic_metrics). Standard-library-only.

- benchmark/grouping.py (new) — the accumulator cluster:
  accumulate_row (renamed from _accumulate_row since it's now the
  module's cross-module entry point) plus private helpers
  _accumulate_calibration, _accumulate_group, _derive_group,
  _ensure_and_accumulate, _update_extreme_list. Imports from
  scoring_primitives; no dependency on scorer.

- benchmark/scorer.py — 2747 → 2222 lines. Keeps CLI, rebuild paths,
  sweep-path score(), cursor bookkeeping. Re-imports the moved
  symbols and calls accumulate_row from grouping.

- benchmark/analyze.py, notify_slack.py — retargeted their direct
  imports of MIN_SAMPLE_SIZE / DISAGREE_THRESHOLD / LARGE_TRADE_THRESHOLD
  to scoring_primitives (would otherwise fail at import time after
  the deletion).

- benchmark/mech_analytics_client.py — docstring refs updated from
  scorer._accumulate_row to grouping.accumulate_row.

- Tests: retargeted direct imports in test_scorer.py, test_analyze.py,
  test_notify_slack.py, test_mech_analytics_client.py. No test logic
  touched. Full suite (1281 tests) passes.

Verification: python3 -c "from benchmark import scorer,
scoring_primitives, grouping" — no circular deps. All six tomte
linters clean on the touched files.
