+++ benchmark/tests/test_scorer.py
@@ -28,43 +28,44 @@ from typing import Any
 from unittest.mock import patch
 
 import pytest
+from benchmark.grouping import _accumulate_group, _derive_group
 from benchmark.scorer import (
-    DISAGREE_THRESHOLD,
-    LARGE_TRADE_THRESHOLD,
-    LATENCY_RESERVOIR_SIZE,
     PLATFORMS,
-    WORST_BEST_SIZE,
-    _accumulate_group,
     _cli_legacy_full_recompute,
     _cli_period,
-    _derive_group,
     _derive_platform_path,
     _derive_tournament_path,
-    _empty_group,
     _is_edge_eligible,
     _partition_rows_by_platform,
     _score_extreme_predictions,
     _score_latency_reservoir,
-    brier_score,
-    classify_difficulty,
-    classify_disagreement,
-    classify_horizon,
-    classify_liquidity,
     compute_calibration_regression,
     compute_ece,
     compute_group_stats,
-    disagree_bucket,
-    edge_score,
     group_by,
     group_by_horizon,
     group_by_month,
     load_history,
-    log_loss_score,
     rebuild,
     score,
     score_period,
     score_period_split_by_platform,
     update,
+)
+from benchmark.scoring_primitives import (
+    DISAGREE_THRESHOLD,
+    LARGE_TRADE_THRESHOLD,
+    LATENCY_RESERVOIR_SIZE,
+    WORST_BEST_SIZE,
+    _empty_group,
+    brier_score,
+    classify_difficulty,
+    classify_disagreement,
+    classify_horizon,
+    classify_liquidity,
+    disagree_bucket,
+    edge_score,
+    log_loss_score,
 )
 
 # ---------------------------------------------------------------------------
